Automatically reset signal handlers on workers under Supervisor
