+++ simpleflow/process/__init__.py
@@ -1,2 +1,2 @@
-from .supervisor import Supervisor
+from .supervisor import Supervisor, reset_signal_handlers
 from .named_mixin import NamedMixin, with_state
